Move authOptions out of API route folder
